[#1025] Added document upload
- Removed ProjectDocument model
- Changed Link model
  - Added IATI fields
  - Added document upload
  - Changed 'caption' to 'title'
- WIP because this migration will conflict with new migrations in other branches
